requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-functions-durable

annotated-types==0.6.0
pydantic==2.6.2
pydantic_core==2.16.3
typing_extensions==4.9.0
